add mod cost
